# stats: Guard regex lookups with substring searches where possible