Annotate addRegex calls with expected substrings.  It is significantly faster to search a candidate stat for these substrings prior to running regex analysis.  This change speeds up 10k case from 8 seconds to 3.5 sec.

Before:
```
Duration(us)  # Calls  Mean(ns)  StdDev(ns)  Min(ns)  Max(ns)  Category  Description                   
     1198153   680000      1761      1408.5     1126  1052984  re-miss   envoy.grpc_bridge_method      
     1184283   680000      1741     1409.55     1119  1048041  re-miss   cipher_suite                  
     1179557   680000      1734     579.372     1112   121057  re-miss   envoy.grpc_bridge_service     
      686269   680127      1009     562.467      483    75670  re-miss   envoy.response_code_class     
      665593   680147       978     411.618      490    79632  re-miss   envoy.response_code           
      616527   680031       906     418.034      517    85152  re-miss   envoy.http_conn_manager_prefix
      451858   680000       664     1333.08      583  1034655  re-match  envoy.cluster_name            
         226      106      2135     999.719      971     9424  re-miss   envoy.dynamo_partition_id     
         215      106      2036     1121.43      915    11952  re-miss   envoy.http_user_agent         
         212      106      2003     580.828      958     3485  re-miss   envoy.dynamo_operation        
         205      106      1935     551.338      928     3387  re-miss   envoy.dyanmo_table            
         202      106      1906     546.486      923     3352  re-miss   envoy.fault_downstream_cluster
          83      116       723     324.772      466     2208  re-match  envoy.http_conn_manager_prefix
          33       14      2373     282.729     2020     2852  re-miss   envoy.ssl_cipher              
          21       20      1056     260.687      651     1439  re-match  envoy.response_code_class     
          16        9      1784     342.155     1515     2663  re-miss   envoy.listener_address        
           4        5       845     347.121      657     1464  re-match  envoy.listener_address        
```

After:
```
Duration(us)  # Calls  Mean(ns)  StdDev(ns)  Min(ns)  Max(ns)  Category        Description                   
      457801   680000       673     859.183      561   198777  re-match        envoy.cluster_name            
      158922   150042      1059     1317.31      707   193774  re-miss         envoy.response_code           
      158747   150022      1058      1240.3      732   202819  re-miss         envoy.response_code_class     
       44221   530105        83     1593.54       33  1140472  re-skip-substr  envoy.response_code           
       42209   680000        62     323.569       41   180224  re-skip-substr  cipher_suite                  
       37968   680000        55     278.834       35   152175  re-skip-substr  envoy.grpc_bridge_method      
       37254   680000        54     346.805       35   172786  re-skip-substr  envoy.grpc_bridge_service     
       35105   530105        66     294.315       30   120865  re-skip-substr  envoy.response_code_class     
          68      116       591     426.393      417     2952  re-match        envoy.http_conn_manager_prefix
          39       14      2803     2468.54     1517    10385  re-miss         envoy.ssl_cipher              
          17       20       890     198.915      622     1182  re-match        envoy.response_code_class     
          10        9      1215     236.915     1005     1785  re-miss         envoy.listener_address        
           6      106        65     22.9534       41      234  re-skip-substr  envoy.dynamo_partition_id     
           6      106        65      14.675       47      143  re-skip-substr  envoy.dynamo_operation        
           6      106        62     13.1069       42      154  re-skip-substr  envoy.fault_downstream_cluster
           6      106        61      12.603       43      124  re-skip-substr  envoy.http_user_agent         
           5      106        49     11.6811       32      130  re-skip-substr  envoy.dyanmo_table            
           4        5       824     309.373      661     1376  re-match        envoy.listener_address        
           0        4        73     21.8689       53      103  re-skip-substr  envoy.http_conn_manager_prefix
```


*Description*:
Allows explicit specification of a required substring, which can be quickly scanned for in an input string before applying regexes.

*Risk Level*: Medium - mis-specifying these can lead to broken tag processing

*Testing*:
//test/...

*Release Notes*: N/A